MODOAIPMH-499 - handle mod-configuration exceptions
