+++ src/main/java/org/folio/oaipmh/helpers/GetOaiRecordHelper.java
@@ -84,7 +84,7 @@ public class GetOaiRecordHelper extends AbstractGetRecordsHelper {
     if (!records.isEmpty()) {
       oaipmh.withGetRecord(new GetRecordType().withRecord(records.iterator().next()));
     } else {
-      oaipmh.withErrors(createNoRecordsFoundError());
+      oaipmh.withErrors(createNoRecordFoundError());
     }
   }
 
@@ -95,8 +95,7 @@ public class GetOaiRecordHelper extends AbstractGetRecordsHelper {
     }
   }
 
-  @Override
-  protected OAIPMHerrorType createNoRecordsFoundError() {
+  private OAIPMHerrorType createNoRecordFoundError() {
     return new OAIPMHerrorType().withCode(ID_DOES_NOT_EXIST).withValue(RECORD_NOT_FOUND_ERROR);
   }
